Extract integration and stress tests according to responsibilities
- Unit test: Test the implemented source code. Write in the package that
 implements the source code.
- Integration test: Test the interface exposed to the user. Usability
 can be checked. Write it in `test/integration`.
- Stress test: Test performance. Write it in `test/stress`.

This commit introduce build constraints as a way to run integration
 tests and unit tests separately.
