+++ test/integration/client_test.go
@@ -1,0 +1,70 @@
+// +build integration
+
+/*
+ * Copyright 2020 The Yorkie Authors. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package integration
+
+import (
+	"context"
+	"github.com/stretchr/testify/assert"
+	"github.com/yorkie-team/yorkie/client"
+	"testing"
+)
+
+func TestClient(t *testing.T) {
+	t.Run("new/close test", func(t *testing.T) {
+		cli, err := client.NewClient(testYorkie.RPCAddr())
+		assert.NoError(t, err)
+
+		defer func() {
+			err := cli.Close()
+			assert.NoError(t, err)
+		}()
+	})
+
+	t.Run("activate/deactivate test", func(t *testing.T) {
+		cli, err := client.NewClient(testYorkie.RPCAddr())
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		defer func() {
+			err := cli.Close()
+			assert.NoError(t, err)
+		}()
+
+		ctx := context.Background()
+
+		err = cli.Activate(ctx)
+		assert.NoError(t, err)
+		assert.True(t, cli.IsActive())
+
+		// Already activated
+		err = cli.Activate(ctx)
+		assert.NoError(t, err)
+		assert.True(t, cli.IsActive())
+
+		err = cli.Deactivate(ctx)
+		assert.NoError(t, err)
+		assert.False(t, cli.IsActive())
+
+		// Already deactivated
+		err = cli.Deactivate(ctx)
+		assert.NoError(t, err)
+		assert.False(t, cli.IsActive())
+	})
+}
