+++ test/integration/text_test.go
@@ -1,0 +1,137 @@
+// +build integration
+
+/*
+ * Copyright 2020 The Yorkie Authors. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package integration
+
+import (
+	"context"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+
+	"github.com/yorkie-team/yorkie/pkg/document"
+	"github.com/yorkie-team/yorkie/pkg/document/proxy"
+	"github.com/yorkie-team/yorkie/test/helper"
+)
+
+func TestText(t *testing.T) {
+	clients := getActivatedClients(t, 2)
+	c1 := clients[0]
+	c2 := clients[1]
+	defer func() {
+		cleanupClients(t, clients)
+	}()
+
+	t.Run("text test", func(t *testing.T) {
+		ctx := context.Background()
+
+		d1 := document.New(helper.Collection, t.Name())
+		err := c1.Attach(ctx, d1)
+		assert.NoError(t, err)
+
+		err = d1.Update(func(root *proxy.ObjectProxy) error {
+			root.SetNewText("k1")
+			return nil
+		}, "set a new text by c1")
+		assert.NoError(t, err)
+		err = c1.Sync(ctx)
+		assert.NoError(t, err)
+
+		d2 := document.New(helper.Collection, t.Name())
+		err = c2.Attach(ctx, d2)
+		assert.NoError(t, err)
+
+		err = d1.Update(func(root *proxy.ObjectProxy) error {
+			root.GetText("k1").Edit(0, 0, "ABCD")
+			return nil
+		}, "edit 0,0 ABCD by c1")
+		assert.NoError(t, err)
+
+		err = d2.Update(func(root *proxy.ObjectProxy) error {
+			root.GetText("k1").Edit(0, 0, "1234")
+			return nil
+		}, "edit 0,0 1234 by c2")
+		assert.NoError(t, err)
+
+		syncClientsThenAssertEqual(t, []clientAndDocPair{{c1, d1}, {c2, d2}})
+
+		err = d1.Update(func(root *proxy.ObjectProxy) error {
+			root.GetText("k1").Edit(2, 3, "XX")
+			return nil
+		}, "edit 2,3 XX by c1")
+		assert.NoError(t, err)
+
+		err = d2.Update(func(root *proxy.ObjectProxy) error {
+			root.GetText("k1").Edit(2, 3, "YY")
+			return nil
+		}, "edit 2,3 YY by c2")
+		assert.NoError(t, err)
+
+		syncClientsThenAssertEqual(t, []clientAndDocPair{{c1, d1}, {c2, d2}})
+
+		err = d1.Update(func(root *proxy.ObjectProxy) error {
+			root.GetText("k1").Edit(4, 5, "ZZ")
+			return nil
+		}, "edit 4,5 ZZ by c1")
+		assert.NoError(t, err)
+
+		err = d2.Update(func(root *proxy.ObjectProxy) error {
+			root.GetText("k1").Edit(2, 3, "TT")
+			return nil
+		}, "edit 2,3 TT by c2")
+		assert.NoError(t, err)
+
+		syncClientsThenAssertEqual(t, []clientAndDocPair{{c1, d1}, {c2, d2}})
+	})
+
+	t.Run("rich text test", func(t *testing.T) {
+		ctx := context.Background()
+
+		d1 := document.New(helper.Collection, t.Name())
+		err := c1.Attach(ctx, d1)
+		assert.NoError(t, err)
+
+		err = d1.Update(func(root *proxy.ObjectProxy) error {
+			root.SetNewRichText("k1").Edit(0, 0, "Hello world", nil)
+			return nil
+		}, `set a new text with "Hello world" by c1`)
+		assert.NoError(t, err)
+		err = c1.Sync(ctx)
+		assert.NoError(t, err)
+
+		d2 := document.New(helper.Collection, t.Name())
+		err = c2.Attach(ctx, d2)
+		assert.NoError(t, err)
+
+		err = d1.Update(func(root *proxy.ObjectProxy) error {
+			text := root.GetRichText("k1")
+			text.SetStyle(0, 1, map[string]string{"b": "1"})
+			return nil
+		}, `set style b to "H" by c1`)
+		assert.NoError(t, err)
+
+		err = d2.Update(func(root *proxy.ObjectProxy) error {
+			text := root.GetRichText("k1")
+			text.SetStyle(0, 5, map[string]string{"i": "1"})
+			return nil
+		}, `set style i to "Hello" by c2`)
+		assert.NoError(t, err)
+
+		syncClientsThenAssertEqual(t, []clientAndDocPair{{c1, d1}, {c2, d2}})
+	})
+}
